fix: amend copyright attributions (#110)
This commit amends copyright attributions that were omitted due to an oversight
on part of the Peerd authors. Copyright header attributions in a few files have been
updated to include "2023 Xenit AB and 2024 The Spegel Authors". The attribution in
the LICENSE file has also been updated to reflect the same.

Fixes #109